requirements.txt
streamlit>=1.32.0
sec-edgar-downloader>=0.4.3
pandas>=2.1.0
plotly>=5.18.0
requests>=2.31.0
lxml>=5.1.0
python-dotenv>=1.0.0
